Exclusively use our own Any trait
