+++ crates/runestick-macros/src/context.rs
@@ -3,18 +3,13 @@ use proc_macro2::Span;
 use proc_macro2::TokenStream;
 use quote::quote;
 use syn::Meta::*;
-use syn::NestedMeta::*;
 
 /// Parsed field attributes.
 #[derive(Default)]
-pub(crate) struct FieldAttrs {
-    /// If the field is marked with `#[rune(any)]`.
-    pub(crate) any: bool,
-}
+pub(crate) struct FieldAttrs {}
 
 pub(crate) struct Context {
     pub(crate) errors: Vec<syn::Error>,
-    pub(crate) any: TokenStream,
     pub(crate) value: TokenStream,
     pub(crate) vm_error: TokenStream,
     pub(crate) vm_error_kind: TokenStream,
@@ -22,7 +17,6 @@ pub(crate) struct Context {
     pub(crate) tuple: TokenStream,
     pub(crate) from_value: TokenStream,
     pub(crate) to_value: TokenStream,
-    pub(crate) from_any: TokenStream,
 }
 
 impl Context {
@@ -30,7 +24,6 @@ impl Context {
     pub fn new() -> Self {
         Self {
             errors: Vec::new(),
-            any: quote!(runestick::Any),
             value: quote!(runestick::Value),
             vm_error: quote!(runestick::VmError),
             vm_error_kind: quote!(runestick::VmErrorKind),
@@ -38,7 +31,6 @@ impl Context {
             tuple: quote!(runestick::Tuple),
             from_value: quote!(runestick::FromValue),
             to_value: quote!(runestick::ToValue),
-            from_any: quote!(runestick::FromAny),
         }
     }
 
@@ -64,14 +56,11 @@ impl Context {
 
     /// Parse field attributes.
     pub(crate) fn parse_field_attrs(&mut self, attrs: &[syn::Attribute]) -> Option<FieldAttrs> {
-        let mut output = FieldAttrs::default();
+        let output = FieldAttrs::default();
 
         for attr in attrs {
             for meta in self.get_rune_meta_items(attr)? {
                 match meta {
-                    Meta(Path(word)) if word == ANY => {
-                        output.any = true;
-                    }
                     meta => {
                         self.errors
                             .push(syn::Error::new_spanned(meta, "unsupported attribute"));
