+++ crates/runestick-macros/tests/test_macro.rs
@@ -11,28 +11,24 @@ struct TestNamed {
     a: OwnedMut<String>,
     b: OwnedMut<Tuple>,
     c: OwnedMut<Object>,
-    #[rune(any)]
     d: OwnedRef<Custom>,
-    #[rune(any)]
     e: OwnedMut<Custom>,
 }
 
 #[derive(FromValue)]
-struct TestUnnamed(OwnedMut<String>, #[rune(any)] OwnedMut<Custom>);
+struct TestUnnamed(OwnedMut<String>, OwnedMut<Custom>);
 
 #[derive(ToValue)]
 struct Test2 {
     a: String,
     b: Tuple,
     c: Object,
-    #[rune(any)]
     d: Custom,
-    #[rune(any)]
     e: Custom,
 }
 
 #[derive(ToValue)]
-struct Test2Unnamed(String, #[rune(any)] Custom);
+struct Test2Unnamed(String, Custom);
 
 #[test]
 fn test_macro() {}
